0.2.4 cmake fix
- glutInitDisplayMode changed to GLUT_RGBA because why not
- added comments to consider freeglut
- adding a sound destructor
- fixed cmake automation, now will use mingw inside the repo directory if there any

+++ tools/cmakeautomate.py
@@ -16,7 +16,9 @@ folder_name = "build"
 # Create the absolute path to the folder
 build_folder_path = os.path.join(script_dir, folder_name)
 
-shutil.rmtree(build_folder_path)
+if os.path.exists(build_folder_path):
+    shutil.rmtree(build_folder_path)
+
 os.makedirs(build_folder_path, exist_ok=True)
 
 # Define the URL to download and the filename
@@ -42,7 +44,20 @@ else:
     print("Download failed.")
     exit()
 
-subprocess.call("cmake -G \"Ninja\" -DCMAKE_MAKE_PROGRAM=ninja.exe ..", shell=True, cwd=build_folder_path)
+ninja_exe_path = os.path.join(build_folder_path, "ninja.exe")
+mingw_path = os.path.join(script_dir, "mingw32")
+if os.path.exists(mingw_path):
+    c_compiler_path = os.path.join(script_dir, r"mingw32\bin\gcc.exe").replace("\\","/")
+    cxx_compiler_path = os.path.join(script_dir, r"mingw32\bin\c++.exe").replace("\\","/")
+    # cmd = ["cmake", "-G", "\"Ninja\"",
+    #        f"-DCMAKE_MAKE_PROGRAM={ninja_exe_path}",
+    #        f"-DCMAKE_C_COMPILER={c_compiler_path}",
+    #        f"-DCMAKE_CXX_COMPILER={cxx_compiler_path}",
+    #        ".."]
+    subprocess.call(f"cmake -G \"Ninja\" -DCMAKE_MAKE_PROGRAM={ninja_exe_path} -DCMAKE_PREFIX_PATH={mingw_path} -DCMAKE_C_COMPILER={c_compiler_path} -DCMAKE_CXX_COMPILER={cxx_compiler_path} ..", shell=True, cwd=build_folder_path)
+else:
+    subprocess.call("cmake -G \"Ninja\" -DCMAKE_MAKE_PROGRAM=ninja.exe ..", shell=True, cwd=build_folder_path)
+    
 subprocess.call("cmake ..", shell=True, cwd=build_folder_path)
 subprocess.call("ninja", shell=True, cwd=build_folder_path)
 print(f"Done building: {build_folder_path}")
